Bluesteroids

+++ Spaceship.pde
@@ -17,16 +17,14 @@ class Spaceship extends Floater
 	    myDirectionX = myDirectionY = 0;
 	    myPointDirection = 0;
     }
-
-	public void setX(int x){myCenterX = x;}  
-	public int getX(){return (int)myCenterX;}   
-	public void setY(int y){myCenterY = y;}   
-	public int getY(){return (int)myCenterY;}   
+	public void setX(double x){myCenterX = x;}  
+	public double getX(){return myCenterX;}   
+	public void setY(double y){myCenterY = y;}   
+	public double getY(){return myCenterY;}   
 	public void setDirectionX(double x){myDirectionX = x;}  
 	public double getDirectionX(){return myDirectionX;}   
 	public void setDirectionY(double y){myDirectionY = y;}  
 	public double getDirectionY(){return myDirectionY;}   
-	public void setPointDirection(int degrees){myPointDirection = degrees;}  
+	public void setPointDirection(double degrees){myPointDirection = degrees;}  
 	public double getPointDirection(){return myPointDirection;}
-
 }
